Before addind Lea's css
