+++ src/DataFixtures/ApplicationFixtures.php
@@ -8,6 +8,7 @@ use Faker\Factory;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 use App\DataFixtures\UserFixtures;
+use App\Entity\User;
 
 class ApplicationFixtures extends Fixture implements DependentFixtureInterface
 {
@@ -20,7 +21,7 @@ class ApplicationFixtures extends Fixture implements DependentFixtureInterface
             $application->setStatus($status)
                 ->setCreatedAt($faker->dateTime())
                 ->setOffer($this->getReference('offer_' . rand(1, 6)))
-                ->setCandidate($this->getReference('user_' . UserFixtures::USER[rand(0, 3)]['Email'])->getCandidate());
+                ->setCandidate($this->getReference('user_' . User::USER[rand(0, 3)]['Email'])->getCandidate());
 
             $manager->persist($application);
         }
