+++ src/DataFixtures/MetadataFixtures.php
@@ -15,7 +15,7 @@ class MetadataFixtures extends Fixture implements DependentFixtureInterface
     public function load(ObjectManager $manager): void
     {
         $faker = Factory::create();
-        foreach (UserFixtures::USER as $user) {
+        foreach (User::USER as $user) {
             if ($user['Role'] === User::ROLE_CANDIDATE) {
                 foreach (CandidateMetadata::METADATA as $type) {
                     $metadata = new CandidateMetadata();
