+++ templates/Offer/show.html.twig
@@ -13,14 +13,16 @@
 		</h3>
 		<section class="otherInfo">
 			<div class="greyAreaInOfferShow">
+				<img src="{{ asset('build/images/picto/location1.c5d41a07.svg') }}" alt="pictoLocation">
 				{{ offer.location }}
 			</div>
 			<div class="greyAreaInOfferShow">
+				<img src="{{ asset('build/images/picto/job 3.2f5799e1.svg') }}" alt="pictoCartable">
 				{{ offer.contract }}
 			</div>
 			<div class="greyAreaInOfferShow">
-				{{ offer.minSalary // 1000 }}K
-																																																																																																																																																																																																																																																																																																																																																																																																																																																																																																																																																																																				-
+				<img src="{{ asset('build/images/picto/piggy-bank 1.00856096.svg') }}" alt="pictoCochon">
+				{{ offer.minSalary // 1000 }}K																																							-
 				{{ offer.maxSalary // 1000 }}K
 			</section>
 			<p>offer posted
@@ -31,27 +33,40 @@
 			</button>
 		</div>
 	</section>
-	<h2 class="mt-4">Company</h2><hr>
+	<h2 class="mt-4">
+		<img src="{{ asset('build/images/picto/immeuble-de-bureaux 2.02421ea4.svg') }}" alt="pictoImmeuble">
+		Company</h2><hr>
 	{{ offer.company.presentation }}
-	<h2 class="mt-4">Job Description</h2><hr>
+	<h2 class="mt-4">
+		<img src="{{ asset('build/images/picto/job 3.2f5799e1.svg') }}" alt="pictoCartable">
+		Job Description</h2><hr>
 	<p>{{ offer.description }}</p>
-	<h2 class="mt-4">Candidate Profil</h2><hr>
-	{% for skill in offer.skills %}
-		<table>
-			{% if skill.type == 'hard' %}
-				<tr>
-					<td>{{ skill.name }}</td>
-				</tr>
-			{% endif %}
-			{% if skill.type == 'soft' %}
-				<tr>
-					<td>{{ skill.name }}</td>
-				</tr>
-			{% endif %}
-		</table>
-	{% endfor %}
-	<h2 class="mt-4">Interview Process</h2><hr>
-	<ol>
+	<h2 class="mt-4">
+		<img src="{{ asset('build/images/picto/description-de-lemploi.224a23d1.svg') }}" alt="pictoPanneau">
+		Candidate Profil</h2><hr>
+	<table class="table">
+		<thead>
+			<tr>
+				<th scope="col">Soft</th>
+				<th scope="col">Hard</th>
+			</tr>
+		</thead>
+		<tbody>
+			{% for skill in offer.skills %}
+				{% if skill.type == 'hard' %}
+					<tr>
+						<th scope="row">{{ skill.name }}</th>
+					{% endif %}
+					{% if skill.type == 'soft' %}
+						<th scope="row">{{ skill.name }}</th>
+					</tr>
+				{% endif %}
+			{% endfor %}
+		</tbody>
+	</table>
+	<h2 class="mt-4">
+		<img src="{{ asset('build/images/picto/interview 2.a6d981fe.svg') }}" alt="pictoItvProcess">
+		Interview Process</h2><hr><ol>
 		<li>
 			{{ offer.interviewProcess }}
 		</li>
